Lots of minor improvements
