+++ tsconfig.json
@@ -8,6 +8,7 @@
 		"strictNullChecks": true,
 		"sourceMap": false,
 		"esModuleInterop": true,
+		"noUnusedLocals": false
 	},
 	"exclude": [
 		"node_modules",
